requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-RESTful==0.3.10
flask-cors == 4.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
supabase==2.13.0